# housekeeping
. add note to HAVE_C99 as per last meeting.
. moved a few extern declarations to header (required for some code analysis tools)
. tiny warning fix in SoundGroup

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4772 3db68df2-c116-0410-a063-a993310a9797

+++ source/sound/SoundGroup.cpp
@@ -89,7 +89,7 @@ void CSoundGroup::PlayNext()
 	{
 		// try loading on the fly only when we need the sound to see if that fixes release problems...
 		if(TestFlag(eRandOrder))
-			m_index = (size_t)rand(0, filenames.size());
+			m_index = (size_t)rand(0, (uint)filenames.size());
 		Handle temp;
 		temp = snd_open(m_filepath + filenames[m_index]);
 		snd_set_gain(temp, m_Gain);
